Add files via upload
There is an error when the player and enemy collide. If the player is standing still and the enemy collides into it, then the enemy goes back. If the enemy is standing still and the player collides into it, then the player goes back. The first error is still present (jumping).