Add files via upload
The error in the previous file (written in the comment) has been solved. Attacking the enemy from behind doesn't cause the enemy to move randomly/weirdly. Collision from front of enemy and behind has been solved. But there is an error where the player can push the enemy out of the game.