Add files via upload
There was a glitch (not noticed earlier) the addition and subtraction of health was looping continuously to infinite values. I stopped this from happening so the max and min health is 30 and 0. I also changed the dimensions that the powerup can spawn in, to prevent it spawning in the walls around the game.  All the powerups now work.